+++ src/types.ts
@@ -1,0 +1,7 @@
+export interface ObjectBindingsPermissionObj {
+	[key: string]: true | ObjectBindingsPermissionObjPerm
+}
+
+export interface ObjectBindingsPermissionObjPerm {
+	[key: string]: string[]
+}